% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{retrieve.ids}
\alias{retrieve.ids}
\title{Retrieve UGent author IDs}
\usage{
retrieve.ids()
}
\value{
a character vector containing all UGent author IDs
}
\description{
Retrieves all the author IDs by parsing the biblio webpage.
}
\examples{
ids <- retrieve.ids()
ids[1:10]
}
