% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{retrieve.person.data}
\alias{retrieve.person.data}
\title{Retrieve meta data of UGent author IDs}
\usage{
retrieve.person.data(ids)
}
\arguments{
\item{ids}{a character vector containing UGent author IDs}
}
\value{
a data frame containing information pertaining the given UGent authors
}
\description{
Retrieves all the meta data of given UGent author IDs by parsing the biblio webpage.
If run for all UGent IDs, this might take a very long time. In order not to lose progress,
it's advised to retrieve all this data in chunks of 100 IDs, and if something
goes wrong to try that chunk again.
}
\examples{
ids <- c("001997743793", "801001497901", "801001188309")
person.data <- retrieve.person.data(ids)
}

