% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{retrieve.publication.data}
\alias{retrieve.publication.data}
\title{Retrieve publications of UGent authors}
\usage{
retrieve.publication.data(ids, mc_cores = 1)
}
\arguments{
\item{ids}{a character vector containing UGent author IDs}

\item{mc_cores}{The number of cores to use, or a qsub qsub_config}
}
\value{
a list containing a publications data frame and a list containing each publication's authors
}
\description{
Retrieves the publications of given UGent author IDs by parsing the biblio webpage.
If run for all UGent IDs, this might take a very long time. In order not to lose progress,
it's advised to retrieve all this data in chunks of 100 IDs, and if something
goes wrong to try that chunk again.
}
\examples{
ids <- c("001997743793", "801001497901", "801001188309")
person.data <- retrieve.publication.data(ids)
}
